fixed styles, changed file name
