--- src/modules/dashboard/components/DAOIncorporation/IncorporationForm/SingleUserPicker/SingleUserPicker.css
@@ -1,35 +1,0 @@
-.protectorWrapper {
-  display: flex;
-  justify-content: space-between;
-  padding: 11px 0 14px;
-  font-size: var(--size-smallish);
-  font-weight: var(--weight-bold);
-  color: color-mod(var(--temp-grey-blue-7) alpha(85%));
-}
-
-.selectWrapper {
-  display: flex;
-  justify-content: flex-start;
-  align-items: center;
-}
-
-.protectorError {
-  font-size: var(--size-tiny);
-  white-space: nowrap;
-  color: var(--danger);
-  letter-spacing: var(--spacing-medium);
-}
-
-.protectorWrapper label {
-  margin-right: 0;
-}
-
-.deleteIcon {
-  height: 20px;
-  width: 20px;
-  fill: var(--pink);
-}
-
-.deleteIcon:hover {
-  cursor: pointer;
-}
