--- src/modules/dashboard/components/DAOIncorporation/IncorporationForm/SingleUserPicker/SingleUserPicker.css.d.ts
@@ -1,4 +1,0 @@
-export const protectorWrapper: string;
-export const selectWrapper: string;
-export const protectorError: string;
-export const deleteIcon: string;
