--- src/modules/dashboard/components/DAOIncorporation/IncorporationForm/SingleUserPicker/index.ts
@@ -1,1 +1,0 @@
-export { default } from './SingleUserPicker';
