--- src/modules/dashboard/components/DAOIncorporation/Stages/StageItem/StageItem.css.d.ts
@@ -1,8 +1,0 @@
-export const container: string;
-export const dot: string;
-export const verticalLine: string;
-export const activeDot: string;
-export const label: string;
-export const activeLabel: string;
-export const description: string;
-export const completedLabel: string;
