--- src/modules/dashboard/components/DAOIncorporation/Stages/StageItem/index.ts
@@ -1,1 +1,0 @@
-export { default } from './StageItem';
