--- src/modules/dashboard/components/DAOIncorporation/Stages/Stages.css
@@ -1,156 +1,0 @@
-@value iconColor: var(--temp-grey-blue-7);
-
-.mainContainer {
-  padding-top: 20px;
-  max-width: 340px;
-}
-
-.stagesText {
-  display: flex;
-  align-items: center;
-  font-size: var(--size-smallish);
-  font-weight: var(--weight-bold);
-  line-height: 18px;
-  color: var(--temp-grey-blue-7);
-}
-
-.statusContainer {
-  display: flex;
-  justify-content: space-between;
-  padding: 10px 0;
-  border-bottom: 1px solid var(--temp-grey-6);
-}
-
-.statusContainer:first-child {
-  padding-top: 0;
-}
-
-.icon svg {
-  height: 18px;
-  width: 18px;
-  fill: var(--temp-grey-blue-7);
-}
-
-.buttonsContainer {
-  display: flex;
-  align-items: center;
-}
-
-.buttonsContainer button:not(:last-of-type) {
-  margin-right: 30px;
-}
-
-.buttonsContainer button {
-  &:disabled {
-    height: 20px;
-    border: none;
-    border-radius: 5px;
-    background-color: var(--colony-light-blue);
-    font-weight: var(--weight-bold);
-    color: var(--temp-grey-blue-7);
-  }
-
-  &:disabled:hover {
-    background-color: var(--colony-light-blue);
-    color: var(--temp-grey-blue-7);
-    cursor: default;
-  }
-}
-
-.iconButton {
-  margin-left: 0;
-  padding: 0;
-  border: none;
-  background-color: transparent;
-}
-
-.iconButton i {
-  display: flex;
-}
-
-.iconButton:hover {
-  cursor: pointer;
-}
-
-.iconButtonDisabled {
-  opacity: 0.5;
-}
-
-.iconWrapper {
-  position: relative;
-}
-
-.iconWrapper:hover {
-  &::after {
-    height: 30px;
-    width: 30px;
-    position: absolute;
-    top: 0;
-    left: 0;
-    border-radius: 50%;
-    background-color: var(--action-secondary);
-    transform: translate(-6px, -6px);
-    content: '';
-    opacity: 0.1;
-  }
-}
-
-.iconButtonDisabled:hover {
-  cursor: not-allowed;
-}
-
-.buttonTooltip {
-  max-width: 192px;
-}
-
-.formStages {
-  position: relative;
-}
-
-@keyframes fadeIn {
-  from {
-    opacity: 0;
-  }
-
-  to {
-    opacity: 1;
-  }
-}
-
-.formStagesMsg {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 21px;
-  width: 100%;
-  position: absolute;
-  top: -14px;
-  border: 1px solid var(--pink);
-  border-radius: 12px;
-  font-size: 11px;
-  font-weight: 400;
-  line-height: 18px;
-  transform: translateY(-100%);
-  letter-spacing: 0.5px;
-  animation-name: fadeIn;
-  animation-duration: 500ms;
-}
-
-.formStagesMsgText {
-  color: var(--pink);
-}
-
-.formStagesMsgAction {
-  margin: 0 0 0 8px;
-  padding: 0;
-  border: 0;
-  background-color: transparent;
-  font-size: inherit;
-  color: var(--action-secondary);
-  appearance: none;
-  cursor: pointer;
-}
-
-.formStagesMsgAction:focus {
-  opacity: 0.5;
-}
