--- src/modules/dashboard/components/DAOIncorporation/Stages/Stages.css.d.ts
@@ -1,15 +1,0 @@
-export const iconColor: string;
-export const mainContainer: string;
-export const stagesText: string;
-export const statusContainer: string;
-export const icon: string;
-export const buttonsContainer: string;
-export const iconButton: string;
-export const iconButtonDisabled: string;
-export const iconWrapper: string;
-export const buttonTooltip: string;
-export const formStages: string;
-export const formStagesMsg: string;
-export const fadeIn: string;
-export const formStagesMsgText: string;
-export const formStagesMsgAction: string;
