--- src/modules/dashboard/components/DAOIncorporation/Stages/StagesButton/StagesButton.css
@@ -1,14 +1,0 @@
-.button {
-  composes: themePrimary main from '~core/Button/Button.css';
-  padding: 0;
-  height: 29px;
-  width: 100px;
-}
-
-.buttonTooltip {
-  max-width: 192px;
-}
-
-.buttonWithTooltip {
-  margin-left: 20px;
-}
