--- src/modules/dashboard/components/DAOIncorporation/Stages/StagesButton/StagesButton.css.d.ts
@@ -1,3 +1,0 @@
-export const button: string;
-export const buttonTooltip: string;
-export const buttonWithTooltip: string;
