--- src/modules/dashboard/components/DAOIncorporation/Stages/StagesButton/index.ts
@@ -1,1 +1,0 @@
-export { default } from './StagesButton';
