--- src/modules/dashboard/components/DAOIncorporation/Stages/index.ts
@@ -1,2 +1,0 @@
-export { default } from './Stages';
-export { default as FormStages } from './FormStages';
