--- src/modules/dashboard/components/Dialogs/StaartApplicationDialog/index.ts
@@ -1,1 +1,0 @@
-export { default } from './StartApplicationDialog';
