+++ src/modules/dashboard/components/Dialogs/StakeApplicationDialog/index.ts
@@ -1,0 +1,1 @@
+export { default } from './StakeApplicationDialog';
